#2842 Moved HTMLMasonLexer to the common
