+++ mason/framework/common/src/main/java/com/perl5/lang/mason2/idea/configuration/VariableDescription.java
