+++ mason/htmlmason/common/src/main/java/com/perl5/lang/htmlmason/idea/configuration/HTMLMasonCustomTag.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2015-2023 Alexandr Evstigneev
+ * Copyright 2015-2025 Alexandr Evstigneev
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
