+++ mason/htmlmason/common/src/main/java/com/perl5/lang/htmlmason/idea/configuration/HTMLMasonSettings.java
@@ -123,7 +123,7 @@ public class HTMLMasonSettings extends AbstractMasonSettings implements Persiste
   }
 
   @Override
-  protected PerlSourceRootType getSourceRootType() {
+  public PerlSourceRootType getSourceRootType() {
     return HTMLMasonSourceRootType.INSTANCE;
   }
